refactor(member): implement userdetails interface
