+++ src/main/java/nextstep/subway/auth/application/UserDetails.java
@@ -1,0 +1,9 @@
+package nextstep.subway.auth.application;
+
+public interface UserDetails {
+    String getEmail();
+
+    String getPassword();
+
+    boolean checkPassword(String password);
+}
