+++ src/main/java/nextstep/subway/auth/application/UserDetailsService.java
@@ -1,0 +1,5 @@
+package nextstep.subway.auth.application;
+
+public interface UserDetailsService {
+    UserDetails loadUserByUsername(String username);
+}
