+++ src/main/java/nextstep/subway/member/application/CustomUserDetailsService.java
@@ -1,12 +1,13 @@
 package nextstep.subway.member.application;
 
+import nextstep.subway.auth.application.UserDetailsService;
 import nextstep.subway.member.domain.LoginMember;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CustomUserDetailsService implements UserDetailsService{
+public class CustomUserDetailsService implements UserDetailsService {
     private MemberRepository memberRepository;
 
     public CustomUserDetailsService(MemberRepository memberRepository) {
