--- src/main/java/nextstep/subway/member/application/UserDetailsService.java
@@ -1,7 +1,0 @@
-package nextstep.subway.member.application;
-
-import nextstep.subway.member.domain.LoginMember;
-
-public interface UserDetailsService {
-    LoginMember loadUserByUsername(String username);
-}
